Moved ExampleSource into a separate file.
